+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## unreleased
+## Unreleased
 
 ### Breaking Changes
 
@@ -8,10 +8,12 @@
 
 ### Fixes
 
+- feat: Add trace information from scope to event capture (#1065)
 - fix: SentryOptions.tracesSampleRate default value (#1069)
 - ref: Discard unfinished spans when capturing transaction (#1066)
 - ref: Make calls to customSamplingContext nonnull (#1061)
-- fix: fix: performance headers (#1079)
+- fix: performance headers (#1079)
+- fix: Release builds in CI (#1076)
 
 ## 7.0.0-beta.0